Fixed typeDefs and added config folder with connection.js.
