+++ server/schemas/typeDefs.js
@@ -1,5 +1,4 @@
 const { gql } = require('apollo-server-express');
-const productSchema = require('./productSchema');
 
 const typeDefs = gql`
   type Profile {
@@ -15,11 +14,26 @@ const typeDefs = gql`
     profile: Profile
   }
 
+  type Product {
+    _id: ID!
+    productType: String 
+    animalType: String
+    color: String
+  }
+
+  input CreateProductInput {
+    productType: String!
+    animalType: String!
+    color: String!
+  }
+
   type Query {
     profiles: [Profile]!
     profile(profileId: ID!): Profile
     # Because we have the context functionality in place to check a JWT and decode its data, we can use a query that will always find and return the logged in user's data
     me: Profile
+    getProduct(_id: ID!): Product
+    getAllProducts: [Product]
   }
 
   type Mutation {
@@ -29,6 +43,7 @@ const typeDefs = gql`
     addSkill(profileId: ID!, skill: String!): Profile
     removeProfile: Profile
     removeSkill(skill: String!): Profile
+    createProduct(input: CreateProductInput!): Product
   }
 `;
